Swap Navigation item "Events" to "Calendar"
If a website visitor wants to discover the Comaking space, they will want to see the usual opening hours. Those are described in the calendar, which is accessed by the navigation item Events.
A visitor would expect "events" to describe special events (yearly meeting, special seminar, etc...), not usual activities.
A calendar is expected to cover both.